Implement line additional fee feature
